First quries CREATE and GET on user, room, movie + exceptions + model mappers + jsonignore + business reqs

+++ src/main/java/com/backend/cinema/exception/UserNotFoundException.java
@@ -1,8 +1,8 @@
 package com.backend.cinema.exception;
 
-public class UserNotFoundException extends RuntimeException{
+public class UserNotFoundException extends RuntimeException {
 
-	public UserNotFoundException(long id) {
-        super("User with id " + id + " doesn't exist ");
-    }
+	public UserNotFoundException(int id) {
+		super("User with id " + id + " doesn't exist ");
+	}
 }
